PHOENIX_V2/requirements.txt
pandas
numpy
requests
pyyaml
python-dotenv
ib_insync
